fix: update delete route middleware for admin access in archive and physical location controllers

+++ routes/api.php
@@ -34,14 +34,14 @@ Route::prefix('v1')->group(function () {
 
             Route::get('/{id}', [ArchiveController::class, 'show']);
             Route::put('/{id}', [ArchiveController::class, 'update'])->middleware('throttle:30,1');
-            Route::delete('/{id}', [ArchiveController::class, 'destroy'])->middleware('throttle:10,1');
+            Route::delete('/{id}', [ArchiveController::class, 'destroy'])->middleware(['throttle:10,1', 'admin']);
             Route::get('/{id}/preview', [ArchiveController::class, 'preview'])->middleware('throttle:60,1');
             Route::get('/{id}/download', [ArchiveController::class, 'download'])->middleware('throttle:30,1');
 
             Route::get('/{id}/physical-locations', [ArchivePhysicalLocationController::class, 'show']);
             Route::post('/{id}/physical-locations', [ArchivePhysicalLocationController::class, 'store'])->middleware('throttle:30,1');
             Route::put('/{id}/physical-locations', [ArchivePhysicalLocationController::class, 'update'])->middleware('throttle:30,1');
-            Route::delete('/{id}/physical-locations', [ArchivePhysicalLocationController::class, 'destroy'])->middleware('throttle:10,1');
+            Route::delete('/{id}/physical-locations', [ArchivePhysicalLocationController::class, 'destroy'])->middleware(['throttle:10,1', 'admin']);
 
             Route::get('/retention/ready', [ArchiveController::class, 'readyForDestruction']);
             Route::patch('/{id}/retention/decide', [ArchiveController::class, 'decideRetention'])->middleware('throttle:10,1');
